修改优化代码，测试传统udp和iocp udp的差距，单收 当并发送小于30的情况下，传统UDP CPU 使用率较低 当并发数大于30的情况下，IOCP UDP的使用了较低

+++ XML/Src/XMLObject.cpp
@@ -68,7 +68,7 @@ bool XMLObject::parseFile(const std::string& file)
 	fseek(fd,0,SEEK_SET);
 
 	char* buffer = new char[totalsize + 100];
-	int readlen = fread(buffer,1,totalsize,fd);
+	size_t readlen = fread(buffer,1,totalsize,fd);
 	buffer[readlen] = 0;
 	fclose(fd);
